Update newsletter form styling
